requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
pycparser==2.22
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
